Moved all legacy code into a deprecation folder and added deprecation notes
