+++ src/deprecated/config/pointsDisplay.ts
@@ -1,3 +1,9 @@
+/**
+ * @deprecated This file is deprecated and should no longer be used.
+ * Do not add new functionality to this file.
+ * This file will be removed in a future version.
+ */
+
 export const POINTS_DISPLAY_CONFIG = {
   name: "$DARE points",
   shortName: "$DARE",
