+++ src/deprecated/data/mockMatches.ts
@@ -1,4 +1,10 @@
-import { Match } from '../types';
+/**
+ * @deprecated This file is deprecated and should no longer be used.
+ * Do not add new functionality to this file.
+ * This file will be removed in a future version.
+ */
+
+import { Match } from '../../types';
 
 // Mock matches that correspond to the bet match IDs in mockBets.ts
 export const MOCK_MATCHES: Match[] = [
