+++ src/deprecated/utils/networkCheck.ts
