+++ src/pages/MatchDetailPage.tsx
@@ -10,7 +10,7 @@ import { Wallet, AlertTriangle, TrendingUp, ArrowUp, ArrowDown, Calendar, Clock,
 import LoadingSpinner from '../components/common/LoadingSpinner';
 import MatchAnalysisCard from '../components/match/MatchAnalysisCard';
 import MatchBettingForm from '../components/match/MatchBettingForm';
-import { formatDecimalOdds, decimalToAmerican } from '../utils/oddsUtils';
+import { formatDecimalOdds, decimalToAmerican } from '../deprecated/utils/oddsUtils';
 import { Match } from '../types';
 import { MatchAnalysis } from '../types';
